requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
PyJWT==0.3.0
SQLAlchemy==0.9.7
Tempita==0.5.2
Werkzeug==0.9.6
alembic==0.6.7
decorator==3.4.0
itsdangerous==0.24
oauthlib==0.7.2
pbr==0.10.0
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.12
wsgiref==0.1.2
